Moved a function to util for reuse.
